Reworked a few things so that we can use Boost 1.32, which is available pre-installed on modern RedHat boxes.  The main change is the introduction of vw::true_type and vw::false_type, which are the same as the Boost versions for Boost 1.33 and up but are different under 1.32.

+++ src/vw/Mosaic/ImageComposite.h
@@ -143,7 +143,7 @@ namespace mosaic {
       typedef Pyramid value_type;
       PyramidGenerator( ImageComposite& composite, int index ) : m_composite(composite), m_index(index) {}
       size_t size() const {
-        return m_composite.sources[m_index].size() * 1.66; // 1.66 = (5/4)*(4/3)
+        return size_t( m_composite.sources[m_index].size() * 1.66 ); // 1.66 = (5/4)*(4/3)
       }
       boost::shared_ptr<value_type> generate() const;
     };
